DHCP: delete config option dnsmasq_dns_server
This field was marked as deprecated In commit
a269541c603f8923b35b7e722f1b8c0ebd42c95a. That was in Kilo
which has provided enough time for admins to addjust to this.

In addition to this the patch sets the default value as [].
If a value is not specified this is None and that should not be
the default list.

DocImpact
UpgradeImpact

Change-Id: Ieaf18ffc9baf7e1caebe9de47017338bebd92c84