Create segment_host mapping after new segment
This adds logic to add segment to host mappings for agents
after a segment is created. The previous logic was doing this
only after an agent initially reported after the server started.
This meant that any segments created after the agent was running
would not get mappings to the agent's host.

Partially-Implements: blueprint routed-networks
Change-Id: Ibcd9964afb1c7ac3926447f90b3fd364fd4835ca