Only update SegmentHostMapping for the given host
Now, when an agent removes a physical network from its configuration,
not only the segments with that physical network will be unbinded
from the host of upated agent. All the segments with that physical
network will be unbinded from all hosts. This is incorrect, the
segments should only be unbinded from the host of updated agent.

Change-Id: Iccca843d1682ac54ec87c3b003a33a0fc5c62205
Closes-bug: #1592463