ml2 lb: do not program arp responder when unused
When arp_responder is not set, the proxy flag is not set on the VXLAN
VTEP interface so no ARP/ND responses are sent. In this (default case)
it is unnecessary to populate the neighbor table on each VxLAN VTEP
interface.

Related-Bug: #1531013
Change-Id: I0fff2228b5b819829edac0bb6597ecb8e5a036ad
(cherry picked from commit 57848f7ba789fcefe712b1b89026cf0cdaf03436)