tests: clean up designate client session mock on test exit
This test was modifying the driver session method without making an
effort to restore the original value after the test case completion.
This resulted in two consequent tests that relied on the method that
were triggered in the same test thread, to trigger a failure for one of
them.

Changed the setup logic for the test class to use mock.patch(..).start()
instead.

Closes-Bug: #1593647
Change-Id: I08be90691b5417025c40c5a18308d820dc7a43d2