Move state reporting initialization to after worker forking
We start state reporting thread before forking child processes. This
leads to possibility of fork in the middle of reporting process. This
can cause child processes to start with corrupted global state. In [0]
it represented itself as metadata agent not being able to make RPC call
because of lock acquired in main process before fork and released after
fork. See bug description for details.

[0] https://review.openstack.org/312393

Partial-Bug: 1594439

Change-Id: Id5079a296bc402c6f5b6cdb2df72811eab5bc6ed