Retry DBDuplicate errors in retry decorator
The MAC duplicate detection logic now expects that the core
plugin catches DBDuplicate entries and converts them into
retry attempts. This assumption was valid for ML2 because it
catches them, but it's not true for anything else, including
unit tests that just use db_base_plugin_v2 directly.

Instead of expecting each plugin to catch and convert
DBDuplicate errors into RetryRequests, this patch just has
the retry logic check for DBDuplicate errors. A DBDuplicate
being raised to the API layer for anything other than a race
that needs to be retried is a bug anyway since it will be
turned into an HTTP 500.

The MAC generation for the unit test that failed in the bug
report defines the 4th octect of the MAC in the test config
so there are only 65k MAC addresses available per network.
The unit test would then proceed to create 13 ports on the
network, which would give us a ~1/5000 chance of a dup mac
in the unit test using db_base_plugin_v2, which was not
catching the DBDuplicate exception.

Closes-Bug: #1594796
Change-Id: I828f529db8c389ba0ab1eaa5f93ca2f5563048a8