Revert "OVS: don't throw KeyError when duplicate VLAN tags exist"
This reverts commit d2508163cfcede20df641239be081fe63c79150b.
This fix wasn't addressing the root cause of why
_restore_local_vlan_map was processing ports from different
networks that had the same VLAN tag. It was just making it so
the agent could start but led to a chance of breaking a set of
ports.

This should be the correct fix:
I9d722fa4fabd467ded44d9cd291a3fa4d1af90f6

Change-Id: I2799da2cfa92171b7ef627225c5ba08a8e642f34