Cap tempest to < 12.1.0
Latest tempest release is incompatible with urllib3 1.14 that is
enforced in gate using upper constraints. Tempest uses chunked= argument
to construct a request, the argument that is introduced in urllib3 1.15+
only.

This patch is not a backport from master because it's not needed there.
On master, we use tempest plugin, and tempest is executed from inside a
venv that contains all master versions of dependencies, including
urllib3.

Sadly, it's not that easy to switch the stable branch to using the same
tempest plugin since that would involve messing with job settings and
tox targets. It's safer and easier to cap for now and hope we won't
trigger requirement conflicts with that down the road. If we do... well,
let's fix it when and if we do. For the time being, we'll procrastinate
and pray.

Depends-On: I3b48b86ffe75f3c12b4d13e0239530f18744ef69
Change-Id: I3f5888e6ed6e5bb930154f65effb7c8c750edd94