DVR: handle floating IP reassociation on the same host
The case when floating IP is reassociated from one VM to another
on the same host without doing disassociation, was not properly handled:
l3 dvr agent only checked that the floating address was present on the
device in router namespace and considered floating IP confiigured correctly.
This patch adds fip_mapping to router object so fixed IP change can be
detected and handled.

Closes-Bug: #1599089
Change-Id: I25c23bb9ad7b9a9b90f225f37417142e2304deb8