Pull stadium projects from governance.o.o in utility script
For housekeeping reasons we have abandon_old_reviews.sh that is
used to purge blocked reviews from Gerrit. The project query
should be extended to all stadium projects. To avoid hard coding,
fetch the list from governance.o.o, parse it and feed it to
the rest of the script.

Change-Id: Iad40d78b7a72a4435382406373bb9f15c0ffc4a7