Catch missing binding record in provision handler
The 'provisioned' event handler in ML2 uses db.get_port which
disables eagerloading. This means that it can get back a record
which may have its binding record deleted by a concurrent process
before it references it.

So we just need to check that the port has a port_binding before
referencing any of its attributes.

Change-Id: I1a939c549a347004cbe4ca3a3d9307fbb71083ba
Closes-Bug: #1599643