Delete default route if no gateway in external net
Currently, when the default route is added to device, there is no
way to delete it. Even if user clear the gateway ip of external
net, there is still a default route present in router namespace.

This patch will check the default route in namespace first. If its
gateway ip is not one of the gateway ips in external net. This
default route will be deleted.

Change-Id: If2ab88b392f3935306d73e3c2de0a35548567464
Closes-Bug: #1520775