Only ports on routed networks are deferred
A port has deferred IP allocation only if it is on a routed network.
Otherwise, it is just a port without any IP addresses.

The unit tests were testing for deferred IP allocation even in
situations where the network was not routed, so this patch fixes them
up.

Change-Id: Ieeab03c31b8f45d63b5044d85ded704d2fe027a9
Partially-Implements: blueprint routed-networks