Emit registry events on subport addition/removal
When a list of subports has been added or removed from a trunk, the
registry should emit a corresponding event and pass the list of added or
removed SubPort objects as the payload.

Trunk agent RPC callbacks can then be subscribed to these events in
order to relay subport information to the agents.

Partially-implements: blueprint vlan-aware-vms
Change-Id: Icb8da23cdd2244da5183058c31d8acef22d78eaa