Use context from_environ to load contexts
The from_environ method loads a context object with all the appropriate
values that are retrieved from auth_token middleware and other oslo
middlewares. This will let oslo add new things to policy enforcement in
future without having to manually change all services.

To do this we need to pass keyword arguments for the context through to
the base class.

Related-Bug: #1602081
Change-Id: I0533c8aee3893449b757f1e35fc89a451ae1720c