Add RouterPort bindings for all HA ports
Adding these bindings allows us to remove the special-cased HA
interface deletion in l3_hamode_db that had the potential for
leaving orphaned ports on failures. With the bindings present,
the interface deletion is handled before the router delete so
a failure to delete the interface will prevent router deletion.

Related-Bug: #1540271
Change-Id: I2de8503742661c18a2ec2c5ade7ec58ea380e749