fullstack: Add hybrid plug support
Whether hybrid plugging should be used for vm or not is part of
vif_details that Neutron API sends on port creation. This patch uses
this information to decide and creates hybrid plug if it should.

I also spotted that we use interface driver. I don't think that this is
what we should be doing for fullstack vms, we should rather simulate how
nova plugs tap into integration bridge. Interface driver is used for l3
and dhcp ports.

Current implementation of hybrid plugging should be in parity with
non-hybrid (i.e. does what interface driver). There will be either
followup to fix it or I'll do it as a part of this patch.

Change-Id: I40cadc6c4f537dab960157dcffba615815b9325f