Enable passive deletes on trunk deletion
Add missing sqlalchemy cascade="all, delete-orphan" on the
trunk-subports relationship to disable child object loading
on a trunk delete operation, otherwise sqlalchemy would
attempt to update the children with a NULL key, and cause
an integrity violation.

Deleting a trunk with subports is allowed, and a test is
added to cover that.

Partially-implements: blueprint vlan-aware-vms

Change-Id: I143c436ce6d2d3a441c26a4b08730981f1ec2973