Mark DBConnectionError as retriable
DBConnectionError can be thrown if the connection
to the database gets closed (e.g. sql restart or
network connectivity failure). This should be part
of our standard list of exceptions to retry so we
are resillient to intermittent connectivity issues.

The olso db wrap db retry decorator has this option
as well[1] but we can't use it because our logic to
detect retriable errors in MultipleException types.

1. https://github.com/openstack/oslo.db/blob/a79479088029e4fa51def91cb36bc652356462b6/oslo_db/api.py#L123

Partial-Bug: #1596075
Change-Id: I26423e306173d956667b9d1e40f07f2b5bc8d805