objects: loading synthetic fields from defined ORM relationships.
Limit the SQL query number when loading the synthetic fields for Neutron
DB objects. If synthetic fields is defined in SQL model as orm.relationship,
the objects are loaded with main object. No need to query for those resources
once more.

It has impact on performance in scale, to have constant number of SQL queries
when number of resources stored in DB is increasing.

Change-Id: Ie54d9f51642edec1f4533062eb8306b78a1b6f20
Partially-Implements: blueprint adopt-oslo-versioned-objects-for-db