Refactoring config options for l2 agent ext opts
Refactoring neutron l2 agent extension config agent_ext_manager
and fdb_population opts to be in neutron/conf/agent so that all the
configuration options for l2 agent extensions reside in a centralized
location. This simplifies the process of looking up the l2 agent
extension config opts and provides an easy way to import.

Change-Id: I4ae842c6cfd62edad9a1d435ffad6bec760a2809
Partial-Bug: #1563069