Use neutron-lib add_validator for registration
As per [1], neutron-lib 0.3.0 provides a public add/get API for
local validator registration/access. This API is preferred over
directly accessing the validators dict module-level attribute
that's done today when adding a new validator and in fact
direct access to the validators dict is deprecated.

This patch changes all neutron's usage of the validators dict
to use the public API.

Closes-Bug: #1584237

[1] https://review.openstack.org/#/c/324090/

Change-Id: I97680b0a8d88cb753199fce7949cfb24b506a0ff