Add scaffolding for trunk plugin/server-side driver integration
This patch introduces a framework and enhancements that allow the trunk
plugin to support multiple trunk drivers that integrate with Neutron core
technologies. A mechanism to validate the deployment configuration is
provided so that the trunk plugin can verify that it can start correctly.
Fail fast is desired to give prompt feedback to the operator in case of
configuration inconsistencies.

The OVS driver is developed as a proof of concept and the Linux Bridge
driver will be added as a follow up patch.

Partially-implements: blueprint vlan-aware-vms

Change-Id: I5f9f8845e4108754748be3b64cce250c06e20f1f