Restore old assert_ping behavior
assert_ping() was changed recently to be async-friendly, but
the change caused the drop of a single packet to throw an
error.  Since occasionally the first packet is lost due to
address resolution (ARP) we can't use it for checking liveness
of an IP.

Restored assert_ping() and moved updated code to
assert_async_ping() since that is a special-case.

(cherry picked from commit ca2aa3ca05aafcd3b7a81b81f8280ed363009f11)
Change-Id: Ibe69417a0d819d4cd87e2f487c08fd126b1024e2