Make revision bump robust to concurrent removals
If there expired objects in the session that are concurrently
deleted by another sesssion, the revision code will get an
ObjectDeletedError when it goes to check if the expired object
is a related object to the object it is handling. So we need
to catch this error an continue on since there is nothing it
needs to do.

Change-Id: Ie6034f36c34099f2646840cbdd569a4398534df5
Closes-Bug: #1611627