Refactoring config options for mech_sriov opts
Refactoring neutron ml2 plugin mech_sriov config opts to be in
neutron/conf/plugins/ml2/drivers/mech_sriov so that all the
configuration options for mech_sriov reside in a centralized
location. This simplifies the process of looking up the mech_sriov
config opts and provides an easy way to import.

Change-Id: I64b243f91f2e637279456b3b06c2998cd89dc076
Partial-Bug: #1563069