Relocate PortBindingPort DB Model
This patch moves PortBindingPort model under neutron/db/models for
consistency.There are no plans to implement an OVO object for the model
at this point in time. Actually, the model should probably eventually
go away from Neutron tree since it's not used by reference plugin,
and is just a remnant from the times when the neutron tree contained
lots of plugins, some of which were using the model.

Change-Id: I5728fea38816b7e500eaa836d5c5dadef51457f7
Partial-Bug: #1597913