objects: add support for per parent type foreign keys
Objects may need to be linked to different parent objects. For example,
we want to expose segment objects both on networks as well as on ports
(through port binding level objects). This change allows to specify a
foreign key per parent object type.

Note: this patch does NOT add support for multiple foreign keys for the
same parent.

Change-Id: Ibad4db8da08d9114ac7765e56fe5483434a567d9
Partially-Implements: blueprint adopt-oslo-versioned-objects-for-db