Fix init method for HasStandardAttributes
Puttting description in the signature of __init__ would make it
consume the first positional argument that was passed in. This
would lead to a difficult to understand traceback from sqlalchemy
when the object was initialized with a positional.

This correctly checks the kwargs for the description and any other
standard attributes rather than leaving it in a positional.

Closes-Bug: #1614897
Change-Id: I264bf5fbfc834d7f5d86a2ad03f5e68a91e2185c