Increase default packet count to 3 in assert_ping
One packet is typically not enough to determine if
an IP is responding since the first could get dropped
during ARP resolution.  Change the default to 3 since
that should be sufficient.

Also, some callers were specifying a timeout of 5
seconds, which isn't all that useful since if the
packet never made it out, waiting longer isn't
going to help.  Let them use the default of 1.

Change-Id: I9c32b7840522fb8e1ce0ea1b2361f5d35e008aae