Implement check_vlan_transparency to return True in SR-IOV mech driver
As the SR-IOV mechanism driver does not impact VLAN transparency
support, this patch overrides the check_vlan_transparency method
from the MechanismDriver base class and returns 'True'.  In this
way, inclusion of the SR-IOV mechanism driver in the list of
configured mechanism_drivers does not prevent the creation of
VLAN transparent networks.

Change-Id: Ie35ce63102e59891d040c01c847936f918e0d6ea
Closes-Bug: #1616480