Refactor _ipam_get_subnets
This method was difficult to reason about. This refactoring preserves the
original behavior but reduces code duplication and gathers all of the error
reporting logic in to one section so that future patches can more easily
optimize it.

Unit tests from these packages cover this area well.

    neutron.tests.unit.extensions.test_segment
    neutron.tests.unit.extensions.test_subnet_service_types

Change-Id: I074f34f3d8bc4135664a94dee11ed01c952d2f3e
Partially-Implements: blueprint service-subnets