objects: use correct object class name in NeutronPrimaryKeyMissing
We were passing cls.__class__ as the object type into the exception
error message. But in case of DeclarativeObject based classes (all
NeutronDbObject classes), the result of it is a message like:

neutron.objects.base.NeutronPrimaryKeyMissing: For class
DeclarativeObject missing primary keys: set([])

Note the class name is DeclarativeObject, not the intented object class
name.

Change-Id: I005512b4edf6e4c68a803f4180983e4a01e4f036