ml2: Add original port to context on _bind_port
During portbinding a new PortContext with the binding information
for the new host gets created. This patch adds the original
port to the newly created context. This has the following effects

- When the port was not bound before, context.original will be set
  set to None on portbinding.

- When the port was bound before (also if binding failed before)
  context.original is set to this port dict on portbinding. This
  happens mostly in the following scenarios:

  - Nova triggers a reschedule of an instance
    Nova reschedules an instance to another host, after spawning an
    instance on the first host failed for some reason. In this case,
    context.original is set to the port with the binding information
    as it was on the failed host.

  - Live Migration
    The port is now available on another host. context.original is
    set to the port and binding information of the source host.

Context.original can now be used in mechanism drivers that need to be
migration aware. This is the case for the macvtap mechanism driver.
The corresponding functionality will be added with patch [1].

[1] https://review.openstack.org/#/c/361301

Partial-Bug: #1550400

Change-Id: I2b60243366505f6e6e4c716d627255a56a4ba486