Check MTU sanity of trunk port subports
The MTU of VLAN subinterfaces must not be greater than
MTU of the trunk port. This commit adds a check that
verifies that this is indeed the case.

This limitation comes from the restriction that Linux kernel imposes on
MTUs of untagged interfaces and tagged subinterfaces of these
interfaces. For details of the current state, see
https://bugzilla.kernel.org/show_bug.cgi?id=13467, namely comments 7
and 8. Once this limitation is overcome, it would be possible to remove
this check.

Partially-Implements: blueprint vlan-aware-vms
Change-Id: Ic11e9a2cde4336e8e1533c56b63ff7aaf2b7a510