Narrow down DBError to DBReferenceError in DHCP action handler
Commit 85ed7017ff22dd24ef7558010e62598483561354 removed the DBError
handling to let the retry decorator do its magic, however the
full implications of this change were not evaluated. As a result,
DBReferenceError (which derives from DBError) is not processed
correctly and that caused a regression of the existing logic.

Rather than bloat the retry's responsibility even further, this
patch partially reverts commit 85ed7017f by narrowing down the
exception handling to DBReferenceError only.

Related-bug: #1618216

Change-Id: Icf4e5e4145dcdcdc710b8e42044467913ed01ec1