Config: change deprecation for max_fixed_ips_per_port to Ocata
Update the configuration help to indicate that this will be
deprecated in Ocata. Patches to provide an alternative solution
have been deferred to Ocata. So gthis will update the help
to ensure that it comes with an alternative solution.

The release note does not need to be updated as it indicates that
the variable will be deprecated in Newton or Ocata.

TrivialFix

Change-Id: I9adb3391574016811b664d024856f2c5010130f0