Added missing translation marker for an error
This error was not translated before, but may be raised in
neutron-sanity-check tool that uses the function. The missing
translation marker was spotted during
I790777ddcbd1b02218b3db54ae3d5c931d72d4fa review. Since it's not a
concern of the referred patch to add the missing marker, this tiny bit
was put in this follow-up.

TrivialFix

Change-Id: I198d773da0611a60b41fd9f762a85d5aa6387673