devref: cleaning up release guidelines for stadium subprojects
Removed guideline on how to pack release tarballs because all releases
are now automated and do the right job.

Changed the example link to PyPI to a /simple/ one that lists all
releases, not just the latest.

Removed the step that requires sending an announcement email because it
should be automated by openstack/releases repo.

Change-Id: I0d57cecd88ef2da19034fe433fc41c1e861aee3f