Use ConfigParser instead of SafeConfigParser
SafeConfigParser supports interpolation on top of ConfigParser in
Python 2, and SafeConfigParser is deprecated in Python 3.2 and log
warning like " DeprecationWarning: The SafeConfigParser class has
been renamed to ConfigParser in Python 3.2. This alias will be removed
in future versions. Use ConfigParser directly instead." So we can use
ConfigParser if we don't need interpolation.

Closes-Bug: #1618666
Change-Id: Ie5d38b86bb2ba3c83c7eb541de0389c6e4dcf654