Capture SubnetNotFound from update_port call
In delete_subnet, we update the remaining ports with fixed_ips
of the IPs remaining from the undeleted subnets. The issue with
this is that two concurrent subnet_delete calls can result in
trying to update the remaining port with a fixed IP from a
subnet already deleted. This results in a SubnetNotFound exception,
which aborts the whole deletion.

This patch captures these exceptions and 'continues' to restart
the iteration and re-query to get the latest fixed IPs of the
ports to automatically update.

Closes-Bug: #1623422
Change-Id: I4e299217eb2aabbe572f24e4328272873369da0f