Fix migration of legacy router to DVR
We have to ensure ML2's create_port method is not called
in a transaction.

This adds a temporary hack to set an attribute on the context
to skip this check to accomodate an L3 code-path that has a port
creation entangled in a transaction.  This attribute will
ultimately be removed once this path is refactored.

Change-Id: I9c41a7848b22c437bedcdd7eb57f7c9da873b06d
Closes-bug: #1619312