Retry setting mac address on new OVS port 10 times
We've seen several times in the gate an OVS add port call
succeed and then have the mac address set fail to find it
seconds later. The vswitch log frequently shows that it
returns milliseconds later. Until we get to the bottom of
it, we should just retry several times before giving up
and raising.

Closes-Bug: #1618987
Partial-Bug: #1623732
Change-Id: Ia73a9be047093c02f61e3e9ce13d98dccd49dfeb