Disable DHCP on agent port removal
The previous logic was just ripping the interface out without
stopping dnsmasq. This would lead to a file handle remaining to the
interface which would cause OVS to completely freak out and assign
the same ofport to multiple ports.

This preserves the behavior introduced in
I40b85033d075562c43ce4d0e68296211b3241197 but just fully disables
DHCP rather than relying on an exception generation to cause the
resync.

Closes-bug: #1624701
Change-Id: Icdd9ac136eeb3707c912853b134dbb58109e6940