Fix linuxbridge trunk subport RPC event handler
The part of the linuxbridge trunk driver that actually triggered
the dataplane wiring for subport changes was incorrectly trying to
fetch the trunk using the trunk ID with a method expecting the port
ID. This would always result in it thinking the trunk wasn't on the
host and returning without wiring.

This fixes it by adding another method to fetch trunks by trunk ID.
All of the other operations in this function already expect trunk IDs.

This was discovered and the fix verified with the scenario test in
the dependent patch.

Change-Id: I25324bc1445cc083799ff54f8508a9505517ce84
Closes-Bug: #1625136